Refactors api integration to separate request for api_key to its own class
Also refactors private class methods to be in its own inherited base library
